feat: Se añade ruta styles

+++ src/modules/style/style.controller.ts
@@ -1,20 +1,15 @@
-import { NextFunction, Request, Response, Router } from "express";
-import {
-  ArtistCreateDto,
-  ArtistGetAllDto,
-  ArtistI,
-  ArtistService,
-  ArtistUpdateDto,
-} from "@artist";
-import { ControllerI } from "@interfaces";
-import { HttpException } from "src/shared/exceptions";
-import { checkAdminToken, validationMiddleware } from "@middlewares";
-import { IdDto, SlugDto } from "@dtos";
+import { NextFunction, Request, Response, Router } from 'express';
+import { StyleCreateDto, StyleI, StyleService, StyleUpdateDto } from '@style';
+import { ControllerI } from '@interfaces';
+import { HttpException } from '@exceptions';
+import { validationMiddleware } from '../../shared/middlewares/validation.middleware';
+import { GetAllDto, IdDto } from '@dtos';
+import { checkAdminToken } from '@middlewares';
 
-export class ArtistController implements ControllerI {
-  path = "/artists";
+export class StyleController implements ControllerI {
+  path = '/styles';
   router = Router();
-  private artistService = new ArtistService();
+  private styleService = new StyleService();
   constructor() {
     this.initializeRoutes();
   }
@@ -22,7 +17,7 @@ export class ArtistController implements ControllerI {
   private initializeRoutes() {
     this.router.post(
       `${this.path}/getAll`,
-      validationMiddleware(ArtistGetAllDto),
+      validationMiddleware(GetAllDto),
       this.getAll
     );
     this.router.post(
@@ -31,31 +26,18 @@ export class ArtistController implements ControllerI {
       this.getOneById
     );
     this.router.post(
-      `${this.path}/getOneBySlug`,
-      validationMiddleware(SlugDto),
-      this.getOneBySlug
-    );
-    this.router.post(
       `${this.path}/create`,
-      validationMiddleware(ArtistCreateDto),
+      validationMiddleware(StyleCreateDto),
       this.create
     );
     this.router.put(
       `${this.path}/update`,
-      validationMiddleware(ArtistUpdateDto),
+      validationMiddleware(StyleUpdateDto),
       checkAdminToken,
       this.update
     );
-    this.router.delete(
-      `${this.path}/one/:id`,
-        checkAdminToken,
-      this.deleteOne
-    );
-    this.router.delete(
-      `${this.path}/all`,
-        checkAdminToken,
-      this.deleteAll
-    );
+    this.router.delete(`${this.path}/one/:id`, checkAdminToken, this.deleteOne);
+    this.router.delete(`${this.path}/all`, checkAdminToken, this.deleteAll);
   }
 
   private getAll = async (
@@ -64,9 +46,9 @@ export class ArtistController implements ControllerI {
     next: NextFunction
   ) => {
     try {
-      const body: ArtistGetAllDto = request.body;
-      const result = await this.artistService.getAll(body);
-      response.status(200).send(result);
+      const body: GetAllDto = request.body;
+      const items = await this.styleService.getAll(body);
+      response.status(200).send(items);
     } catch (error) {
       next(new HttpException(400, error.message, request, response));
     }
@@ -79,21 +61,7 @@ export class ArtistController implements ControllerI {
   ) => {
     try {
       const body: IdDto = request.body;
-      const result: ArtistI = await this.artistService.getOneById(body);
-      response.status(200).send(result);
-    } catch (error) {
-      next(new HttpException(400, error.message, request, response));
-    }
-  };
-
-  private getOneBySlug = async (
-    request: Request,
-    response: Response,
-    next: NextFunction
-  ) => {
-    try {
-      const body: SlugDto = request.body;
-      const result: ArtistI = await this.artistService.getOneBySlug(body);
+      const result: StyleI = await this.styleService.getOneById(body);
       response.status(200).send(result);
     } catch (error) {
       next(new HttpException(400, error.message, request, response));
@@ -107,7 +75,7 @@ export class ArtistController implements ControllerI {
   ) => {
     try {
       const body: any = request.body;
-      const result = await this.artistService.create(body);
+      const result = await this.styleService.create(body);
       response.status(200).send(result);
     } catch (error) {
       next(new HttpException(400, error.message, request, response));
@@ -121,7 +89,7 @@ export class ArtistController implements ControllerI {
   ) => {
     try {
       const body = request.body;
-      const result = await this.artistService.update(body);
+      const result = await this.styleService.update(body);
       response.status(200).send(result);
     } catch (error) {
       next(new HttpException(400, error.message, request, response));
@@ -135,7 +103,7 @@ export class ArtistController implements ControllerI {
   ) => {
     try {
       const id = request.params.id;
-      const result = await this.artistService.deleteOne(id);
+      const result = await this.styleService.deleteOne(id);
       response.status(200).send(result);
     } catch (error) {
       next(new HttpException(400, error.message, request, response));
@@ -148,7 +116,7 @@ export class ArtistController implements ControllerI {
     next: NextFunction
   ) => {
     try {
-      const result = await this.artistService.deleteAll();
+      const result = await this.styleService.deleteAll();
       response.status(200).send(result);
     } catch (error) {
       next(new HttpException(400, error.message, request, response));
